Move ValidatorsStatuses into base
