+++ consensus/state_processing/src/per_epoch_processing.rs
@@ -1,7 +1,6 @@
 #![deny(clippy::wildcard_imports)]
 
 // FIXME(altair): refactor to remove phase0/base structs, including `EpochProcessingSummary`
-pub use base::{TotalBalances, ValidatorStatus, ValidatorStatuses};
 pub use epoch_processing_summary::EpochProcessingSummary;
 use errors::EpochProcessingError as Error;
 pub use registry_updates::process_registry_updates;
@@ -20,7 +19,6 @@ pub mod registry_updates;
 pub mod resets;
 pub mod slashings;
 pub mod tests;
-pub mod validator_statuses;
 pub mod weigh_justification_and_finalization;
 
 /// Performs per-epoch processing on some BeaconState.
