+++ consensus/state_processing/src/per_epoch_processing/base.rs
@@ -1,7 +1,4 @@
 use super::{process_registry_updates, process_slashings, EpochProcessingSummary, Error};
-pub use crate::per_epoch_processing::validator_statuses::{
-    TotalBalances, ValidatorStatus, ValidatorStatuses,
-};
 use crate::per_epoch_processing::{
     effective_balance_updates::process_effective_balance_updates,
     historical_roots_update::process_historical_roots_update,
@@ -11,10 +8,12 @@ pub use justification_and_finalization::process_justification_and_finalization;
 pub use participation_record_updates::process_participation_record_updates;
 pub use rewards_and_penalties::process_rewards_and_penalties;
 use types::{BeaconState, ChainSpec, EthSpec, RelativeEpoch};
+pub use validator_statuses::{TotalBalances, ValidatorStatus, ValidatorStatuses};
 
 pub mod justification_and_finalization;
 pub mod participation_record_updates;
 pub mod rewards_and_penalties;
+pub mod validator_statuses;
 
 pub fn process_epoch<T: EthSpec>(
     state: &mut BeaconState<T>,
