+++ consensus/state_processing/src/per_epoch_processing/epoch_processing_summary.rs
@@ -1,7 +1,6 @@
 use super::{
     altair::{participation_cache::Error as ParticipationCacheError, ParticipationCache},
-    base::{TotalBalances, ValidatorStatus},
-    validator_statuses::InclusionInfo,
+    base::{validator_statuses::InclusionInfo, TotalBalances, ValidatorStatus},
 };
 use crate::metrics;
