+++ testing/ef_tests/src/cases/epoch_processing.rs
@@ -5,7 +5,6 @@ use crate::decode::{ssz_decode_state, yaml_decode_file};
 use crate::type_name;
 use crate::type_name::TypeName;
 use serde_derive::Deserialize;
-use state_processing::per_epoch_processing::validator_statuses::ValidatorStatuses;
 use state_processing::per_epoch_processing::{
     altair, base,
     effective_balance_updates::process_effective_balance_updates,
@@ -87,7 +86,7 @@ impl<E: EthSpec> EpochTransition<E> for JustificationAndFinalization {
     fn run(state: &mut BeaconState<E>, spec: &ChainSpec) -> Result<(), EpochProcessingError> {
         match state {
             BeaconState::Base(_) => {
-                let mut validator_statuses = ValidatorStatuses::new(state, spec)?;
+                let mut validator_statuses = base::ValidatorStatuses::new(state, spec)?;
                 validator_statuses.process_attestations(state)?;
                 base::process_justification_and_finalization(
                     state,
@@ -107,7 +106,7 @@ impl<E: EthSpec> EpochTransition<E> for RewardsAndPenalties {
     fn run(state: &mut BeaconState<E>, spec: &ChainSpec) -> Result<(), EpochProcessingError> {
         match state {
             BeaconState::Base(_) => {
-                let mut validator_statuses = ValidatorStatuses::new(state, spec)?;
+                let mut validator_statuses = base::ValidatorStatuses::new(state, spec)?;
                 validator_statuses.process_attestations(state)?;
                 base::process_rewards_and_penalties(state, &mut validator_statuses, spec)
             }
@@ -130,7 +129,7 @@ impl<E: EthSpec> EpochTransition<E> for Slashings {
     fn run(state: &mut BeaconState<E>, spec: &ChainSpec) -> Result<(), EpochProcessingError> {
         match state {
             BeaconState::Base(_) => {
-                let mut validator_statuses = ValidatorStatuses::new(&state, spec)?;
+                let mut validator_statuses = base::ValidatorStatuses::new(&state, spec)?;
                 validator_statuses.process_attestations(&state)?;
                 process_slashings(
                     state,
